Added nerualNetwork initialization, but it needed to be debuged

+++ NeuralNet/Data/Image.cpp
@@ -8,7 +8,7 @@ Image::Image()
 	dev_image = nullptr;
 }
 
-Image::Image(float* image, int sizeX, int sizeY, int channel, bool inHost)
+Image::Image(double* image, int sizeX, int sizeY, int channel, bool inHost)
 {
 	loadImage(image, sizeX, sizeY, channel, inHost);
 }
@@ -38,7 +38,7 @@ void Image::freeMemory()
 
 /*
 * »тоговое изображение сохран€етс€ в пам€ть видеокарты с
-* типом float, а также нормализуетс€ в границах от 0 до 1.
+* типом double, а также нормализуетс€ в границах от 0 до 1.
 */
 void Image::loadImage(const char* path, int channel)
 {
@@ -51,22 +51,22 @@ void Image::loadImage(const char* path, int channel)
 	else
 		chImage = SOIL_load_image(path, &sizeX, &sizeY, &this->channel, channel);
 
-	float* image = new float[sizeX * sizeY * this->channel];
+	double* image = new double[sizeX * sizeY * this->channel];
 	for (int x = 0; x < sizeX * sizeY * this->channel; x++)
 		image[x] = chImage[x] / 255.0f;
 
 	delete[sizeX * sizeY * this->channel] chImage;
 
-	cudaError_t error = cudaMalloc(&dev_image, sizeof(float) * sizeX * sizeY * this->channel);
+	cudaError_t error = cudaMalloc(&dev_image, sizeof(double) * sizeX * sizeY * this->channel);
 	if (error != 0) throw;
 
-	error = cudaMemcpy(dev_image, image, sizeof(float) * sizeX * sizeY * this->channel, cudaMemcpyHostToDevice);
+	error = cudaMemcpy(dev_image, image, sizeof(double) * sizeX * sizeY * this->channel, cudaMemcpyHostToDevice);
 	if (error != 0) throw;
 
 	delete[sizeX * sizeY * this->channel] image;
 }
 
-void Image::loadImage(float* image, int sizeX, int sizeY, int channel, bool inHost)
+void Image::loadImage(double* image, int sizeX, int sizeY, int channel, bool inHost)
 {
 	if (dev_image != nullptr)
 		freeMemory();
@@ -75,10 +75,10 @@ void Image::loadImage(float* image, int sizeX, int sizeY, int channel, bool inHo
 		this->sizeX = sizeX;
 		this->sizeY = sizeY;
 		this->channel = channel;
-		cudaError_t error = cudaMalloc(&dev_image, sizeof(float) * sizeX * sizeY * channel);
+		cudaError_t error = cudaMalloc(&dev_image, sizeof(double) * sizeX * sizeY * channel);
 		if (error != 0) throw;
 
-		error = cudaMemcpy(dev_image, image, sizeof(float) * sizeX * sizeY * channel, cudaMemcpyHostToDevice);
+		error = cudaMemcpy(dev_image, image, sizeof(double) * sizeX * sizeY * channel, cudaMemcpyHostToDevice);
 		if (error != 0) throw;
 	}
 	else
@@ -98,36 +98,36 @@ void Image::loadImage(unsigned char* image, int sizeX, int sizeY, int channel)
 	this->sizeY = sizeY;
 	this->channel = channel;
 
-	float* fimage = new float[sizeX * sizeY * channel];
+	double* fimage = new double[sizeX * sizeY * channel];
 	for (int x = 0; x < sizeX * sizeY * channel; x++)
 		fimage[x] = image[x] / 255.0f;
 
-	cudaError_t error = cudaMalloc(&dev_image, sizeof(float) * sizeX * sizeY * channel);
+	cudaError_t error = cudaMalloc(&dev_image, sizeof(double) * sizeX * sizeY * channel);
 	if (error != 0) throw;
 
-	error = cudaMemcpy(dev_image, fimage, sizeof(float) * sizeX * sizeY * channel, cudaMemcpyHostToDevice);
+	error = cudaMemcpy(dev_image, fimage, sizeof(double) * sizeX * sizeY * channel, cudaMemcpyHostToDevice);
 	if (error != 0) throw;
 
 	delete[sizeX * sizeY * channel] fimage;
 }
 
-void Image::getImage(float*& outImage)
+void Image::getImage(double*& outImage)
 {
 	if (outImage == nullptr)
-		outImage = new float[sizeX * sizeY * channel];
-	cudaError_t error = cudaMemcpy(outImage, dev_image, sizeof(float) * sizeX * sizeY * channel, cudaMemcpyDeviceToHost);
+		outImage = new double[sizeX * sizeY * channel];
+	cudaError_t error = cudaMemcpy(outImage, dev_image, sizeof(double) * sizeX * sizeY * channel, cudaMemcpyDeviceToHost);
 	if (error != 0) throw;
 }
 
 void Image::getImage(unsigned char*& outImage)
 {
-	float* fImage = nullptr;
+	double* fImage = nullptr;
 	getImage(fImage);
 	flToCh(fImage, outImage, sizeX, sizeY, channel);
 	delete[sizeX * sizeY * channel] fImage;
 }
 
-void Image::flToCh(float* fimage, unsigned char*& outChImage, int sizeX, int sizeY, int channel)
+void Image::flToCh(double* fimage, unsigned char*& outChImage, int sizeX, int sizeY, int channel)
 {
 	if (outChImage == nullptr)
 		outChImage = new unsigned char[sizeX * sizeY * channel];
@@ -135,10 +135,10 @@ void Image::flToCh(float* fimage, unsigned char*& outChImage, int sizeX, int siz
 		outChImage[x] = fimage[x] * 255;
 }
 
-void Image::chToFl(unsigned char* chImage, float*& outFImage, int sizeX, int sizeY, int channel)
+void Image::chToFl(unsigned char* chImage, double*& outFImage, int sizeX, int sizeY, int channel)
 {
 	if (outFImage == nullptr)
-		outFImage = new float[sizeX * sizeY * channel];
+		outFImage = new double[sizeX * sizeY * channel];
 	for (int x = 0; x < sizeX * sizeY * channel; x++)
 		outFImage[x] = chImage[x] / 255.0f;
 }
