+++ NeuralNet/Data/Samples/ISample.h
@@ -5,6 +5,6 @@ class ISample
 public:
 	
 	virtual void realese() = 0;
-	virtual float* getInputData(int id) = 0;
-	virtual float* getOutputData(int id) = 0;
+	virtual double* getInputData(int id) = 0;
+	virtual double* getOutputData(int id) = 0;
 };
